requirements.txt
black==21.5b1
ipython==7.23.1
pipdeptree==2.0.0
pytorch-lightning==1.3.1
torchvision==0.9.1